[chore] 커뮤니티 폴더 구조 변경 - 게시글과 댓글, 검색 기록으로 분리
