+++ src/main/java/com/wilo/server/community/dto/comment/CommunityPostAuthorDto.java
@@ -1,4 +1,4 @@
-package com.wilo.server.community.dto;
+package com.wilo.server.community.dto.comment;
 
 import com.wilo.server.user.entity.User;
