+++ src/main/java/com/wilo/server/community/dto/comment/CommunityUserCommentListResponseDto.java
@@ -1,4 +1,4 @@
-package com.wilo.server.community.dto;
+package com.wilo.server.community.dto.comment;
 
 import java.util.List;
