+++ src/main/java/com/wilo/server/community/dto/comment/CommunityUserCommentSummaryDto.java
@@ -1,4 +1,4 @@
-package com.wilo.server.community.dto;
+package com.wilo.server.community.dto.comment;
 
 import java.time.LocalDateTime;
