+++ src/main/java/com/wilo/server/community/dto/post/CommunityLikeResponseDto.java
@@ -1,4 +1,4 @@
-package com.wilo.server.community.dto;
+package com.wilo.server.community.dto.post;
 
 public record CommunityLikeResponseDto(
         boolean liked,
