+++ src/main/java/com/wilo/server/community/dto/post/CommunityPostListResponseDto.java
@@ -1,4 +1,4 @@
-package com.wilo.server.community.dto;
+package com.wilo.server.community.dto.post;
 
 import java.util.List;
