+++ src/main/java/com/wilo/server/community/dto/post/CommunityPostSummaryDto.java
@@ -1,6 +1,6 @@
-package com.wilo.server.community.dto;
+package com.wilo.server.community.dto.post;
 
-import com.wilo.server.community.entity.CommunityCategory;
+import com.wilo.server.community.entity.post.CommunityCategory;
 import java.time.LocalDateTime;
 
 public record CommunityPostSummaryDto(
