+++ src/main/java/com/wilo/server/community/dto/search/CommunitySearchHistoryListResponseDto.java
@@ -1,4 +1,4 @@
-package com.wilo.server.community.dto;
+package com.wilo.server.community.dto.search;
 
 import java.util.List;
