+++ src/main/java/com/wilo/server/community/entity/post/CommunityPostImage.java
@@ -1,4 +1,4 @@
-package com.wilo.server.community.entity;
+package com.wilo.server.community.entity.post;
 
 import com.wilo.server.global.entity.BaseEntity;
 import jakarta.persistence.*;
