+++ src/main/java/com/wilo/server/community/repository/CommunityCommentRepository.java
@@ -1,6 +1,6 @@
 package com.wilo.server.community.repository;
 
-import com.wilo.server.community.entity.CommunityComment;
+import com.wilo.server.community.entity.comment.CommunityComment;
 import com.wilo.server.community.repository.query.CommunityCommentQueryRepository;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
