+++ src/main/java/com/wilo/server/community/repository/query/CommunityCommentQueryRepository.java
@@ -1,6 +1,6 @@
 package com.wilo.server.community.repository.query;
 
-import com.wilo.server.community.entity.CommunityComment;
+import com.wilo.server.community.entity.comment.CommunityComment;
 import java.time.LocalDateTime;
 import java.util.List;
 import org.springframework.data.domain.Pageable;
