+++ src/main/java/com/wilo/server/community/repository/query/CommunityPostLikeQueryRepository.java
@@ -1,6 +1,6 @@
 package com.wilo.server.community.repository.query;
 
-import com.wilo.server.community.entity.CommunityPostLike;
+import com.wilo.server.community.entity.post.CommunityPostLike;
 import java.time.LocalDateTime;
 import java.util.List;
 import org.springframework.data.domain.Pageable;
